Move rosbag_saver into test utilities
